+++ src/Tabby/Core/Time/Timestep.cpp
@@ -1,7 +1,0 @@
-#include "Timestep.h"
-
-namespace Tabby {
-
-// Timestep* Timestep::s_Instance = nullptr;
-
-}