+++ src/Tabby/Utils/PlatformUtils.h
@@ -11,18 +11,4 @@ public:
     static std::string SaveFile(const char* filter);
 };
 
-class Time {
-public:
-    static double GetTime();
-    static double GetDeltaTime() { return m_DeltaTime; }
-
-private:
-    static void SetDeltaTime(double deltaTime) { m_DeltaTime = deltaTime; }
-
-private:
-    inline static double m_DeltaTime;
-
-    friend class Application;
-};
-
 }