+++ src/Tabby/World/Components.cpp
@@ -3,12 +3,10 @@
 #include <Tabby/Physics/2D/Physics2DTypes.h>
 #include <Tabby/Physics/2D/Physics2DUtil.h>
 #include <Tabby/Math/Math.h>
-#include <Tabby/Renderer/GLTF.h>
 
 #include <box2d/box2d.h>
-#include "glm/fwd.hpp"
-#include <glm/glm.hpp>
-#include <glm/gtc/matrix_transform.hpp>
+// #include <glm/glm.hpp>
+// #include <glm/gtc/matrix_transform.hpp>
 #define GLM_ENABLE_EXPERIMENTAL
 #include <glm/gtx/matrix_decompose.hpp>
 #include <glm/gtx/quaternion.hpp>
@@ -16,13 +14,13 @@
 namespace Tabby {
 
 // ----- TransformComponent ---------------------
-void TransformComponent::ApplyTransform(const glm::mat4& transform)
+void TransformComponent::ApplyTransform(const Matrix4& transform)
 {
 
     // using namespace glm;
     using T = float;
 
-    glm::mat4 LocalMatrix(transform);
+    Matrix4 LocalMatrix(transform);
 
     // Normalize the matrix.
     if (glm::epsilonEqual(LocalMatrix[3][3], static_cast<float>(0), glm::epsilon<T>()))
@@ -36,10 +34,10 @@ void TransformComponent::ApplyTransform(const glm::mat4& transform)
     }
 
     // Next take care of translation (easy).
-    Translation = glm::vec3(LocalMatrix[3]);
-    LocalMatrix[3] = glm::vec4(0, 0, 0, LocalMatrix[3].w);
+    Translation = Vector3(LocalMatrix[3]);
+    LocalMatrix[3] = Vector4(0, 0, 0, LocalMatrix[3].w);
 
-    glm::vec3 Row[3], Pdum3;
+    Vector3 Row[3], Pdum3;
 
     // Now get scale and shear.
     for (glm::length_t i = 0; i < 3; ++i)
@@ -83,6 +81,73 @@ void TransformComponent::ApplyTransform(const glm::mat4& transform)
     Rotation.z *= Math::RAD2DEG;
 }
 
+void TransformComponent::ApplyTransformToLocal(const Matrix4& transform)
+{
+
+    // using namespace glm;
+    using T = float;
+
+    Matrix4 LocalMatrix(transform);
+
+    // Normalize the matrix.
+    if (glm::epsilonEqual(LocalMatrix[3][3], static_cast<float>(0), glm::epsilon<T>()))
+        return;
+
+    // First, isolate perspective.  This is the messiest.
+    if (glm::epsilonNotEqual(LocalMatrix[0][3], static_cast<T>(0), glm::epsilon<T>()) || glm::epsilonNotEqual(LocalMatrix[1][3], static_cast<T>(0), glm::epsilon<T>()) || glm::epsilonNotEqual(LocalMatrix[2][3], static_cast<T>(0), glm::epsilon<T>())) {
+        // Clear the perspective partition
+        LocalMatrix[0][3] = LocalMatrix[1][3] = LocalMatrix[2][3] = static_cast<T>(0);
+        LocalMatrix[3][3] = static_cast<T>(1);
+    }
+
+    // Next take care of translation (easy).
+    LocalTranslation = Vector3(LocalMatrix[3]);
+    LocalMatrix[3] = Vector4(0, 0, 0, LocalMatrix[3].w);
+
+    Vector3 Row[3], Pdum3;
+
+    // Now get scale and shear.
+    for (glm::length_t i = 0; i < 3; ++i)
+        for (glm::length_t j = 0; j < 3; ++j)
+            Row[i][j] = LocalMatrix[i][j];
+
+    // Compute X scale factor and normalize first row.
+    LocalScale.x = length(Row[0]);
+    Row[0] = glm::detail::scale(Row[0], static_cast<T>(1));
+    LocalScale.y = length(Row[1]);
+    Row[1] = glm::detail::scale(Row[1], static_cast<T>(1));
+    LocalScale.z = length(Row[2]);
+    Row[2] = glm::detail::scale(Row[2], static_cast<T>(1));
+
+    // At this point, the matrix (in rows[]) is orthonormal.
+    // Check for a coordinate system flip.  If the determinant
+    // is -1, then negate the matrix and the scaling factors.
+#if 0
+        		Pdum3 = cross(Row[1], Row[2]); // v3Cross(row[1], row[2], Pdum3);
+        		if (dot(Row[0], Pdum3) < 0)
+        		{
+        			for (length_t i = 0; i < 3; i++)
+        			{
+        				scale[i] *= static_cast<T>(-1);
+        				Row[i] *= static_cast<T>(-1);
+        			}
+        		}
+#endif
+
+    LocalRotation.y = asin(-Row[0][2]);
+    if (cos(Rotation.y) != 0) {
+        LocalRotation.x = atan2(Row[1][2], Row[2][2]);
+        LocalRotation.z = atan2(Row[0][1], Row[0][0]);
+    } else {
+        LocalRotation.x = atan2(-Row[2][0], Row[1][1]);
+        LocalRotation.z = 0;
+    }
+
+    LocalRotation.x *= Math::RAD2DEG;
+    LocalRotation.y *= Math::RAD2DEG;
+    LocalRotation.z *= Math::RAD2DEG;
+}
+
 // --------------------------------------------------
 //
 //
@@ -103,7 +168,7 @@ void Rigidbody2DComponent::SetFixedRotation(bool enable)
     FixedRotation = enable;
 }
 
-void Rigidbody2DComponent::SetVelocity(glm::vec2 velocity)
+void Rigidbody2DComponent::SetVelocity(Vector2 velocity)
 {
     if (B2_IS_NON_NULL(RuntimeBodyId))
         b2Body_SetLinearVelocity(RuntimeBodyId, { velocity.x, velocity.y });
@@ -120,7 +185,7 @@ Rigidbody2DComponent::BodyType Rigidbody2DComponent::GetType() const
     return Type;
 }
 
-glm::vec2 Rigidbody2DComponent::GetVelocity() const
+Vector2 Rigidbody2DComponent::GetVelocity() const
 {
     if (B2_IS_NON_NULL(RuntimeBodyId)) {
         b2Vec2 velocity = b2Body_GetLinearVelocity(RuntimeBodyId);
@@ -137,7 +202,7 @@ float Rigidbody2DComponent::GetAngularVelocity() const
         return 0.0f;
 }
 
-glm::vec2 Rigidbody2DComponent::GetPosition() const
+Vector2 Rigidbody2DComponent::GetPosition() const
 {
     if (B2_IS_NON_NULL(RuntimeBodyId)) {
         b2Vec2 position = b2Body_GetPosition(RuntimeBodyId);
@@ -151,15 +216,12 @@ float Rigidbody2DComponent::GetAngle() const
     float angleRadians;
 
     if (B2_IS_NON_NULL(RuntimeBodyId))
-        angleRadians = b2Body_GetAngle(RuntimeBodyId);
+        angleRadians = b2Rot_GetAngle(b2Body_GetRotation(RuntimeBodyId));
 
-    // Convert the angle from radians to degrees
-    float angleDegrees = angleRadians * (180.0f / b2_pi);
+    float angleDegrees = angleRadians * Math::RAD2DEG;
 
-    // Wrap the angle to the range [0, 360]
     angleDegrees = std::fmod(angleDegrees, 360.0f);
 
-    // If the angle is negative, add 360 to make it positive
     if (angleDegrees < 0.0f) {
         angleDegrees += 360.0f;
     }